refactored implementation -> distribution (to avoid confusion with HLS impl)
