+++ src/current_dist.h
@@ -1,0 +1,14 @@
+#ifndef current_dist_h
+#define current_dist_h
+
+// Include definitions required by distribution files. Fot the sake of
+// simplicity the genreated headers do not contain any project specific type
+// definitions.
+#include "definitions.h"
+#include "obj.h"
+#include "templates.h"
+
+// Includes menu distribution from symbolic link `current_dist`.
+#include "../current_dist/hls/module_0/src/impl/logic.hxx"
+
+#endif
