--- templates/hls/conditions.h
@@ -1,14 +1,0 @@
-/*
- * auto generated menu implementation
- */
-
-#ifndef impl_conditions_h
-#define impl_conditions_h
-
-#include "../../impl/hls/cuts.h"
-
-namespace impl {
-
-} // namespace impl
-
-#endif
