--- templates/hls/cuts.h
@@ -1,16 +1,0 @@
-/*
- * auto generated menu implementation
- */
-
-#ifndef impl_cuts_h
-#define impl_cuts_h
-
-#include "../../src/obj.h"
-
-namespace impl {
-
-{{ cuts }}
-
-} // namespace impl
-
-#endif
