+++ templates/hls/header.hxx
@@ -1,0 +1,10 @@
+/*
+ * auto generated by utils/implement.py
+ *
+ * created: {{ header_timestamp }}
+ * implements: {{ header_menu_name }}
+ *
+ */
+
+{% block code %}
+{% endblock code %}
